Remove the need for two different EditorApp classes
