+++ packages/client/src/vscode/services.ts
@@ -36,6 +36,7 @@ export interface ViewsConfig {
 export interface VscodeApiConfig {
     enableTextmate: boolean;
     vscodeApiInitPerformExternally?: boolean;
+    loadThemes?: boolean;
     serviceOverrides?: monaco.editor.IEditorOverrideServices;
     enableExtHostWorker?: boolean;
     workspaceConfig?: IWorkbenchConstructionOptions;
