+++ packages/wrapper/src/index.ts
@@ -3,8 +3,6 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-export * from './editorAppBase.js';
-export * from './editorAppClassic.js';
-export * from './editorAppExtended.js';
+export * from './editorApp.js';
 export * from './languageClientWrapper.js';
 export * from './wrapper.js';
